pyproject.toml
[project]
name = "trivia-2-the-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asgiref==3.8.1",
    "bcrypt==4.2.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==43.0.3",
    "decorator==5.1.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.4",
    "fastapi-cli==0.0.5",
    "gssapi==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "jinja2==3.1.4",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "passlib==1.7.4",
    "psycopg==3.2.3",
    "psycopg-binary==3.2.3",
    "psycopg-pool==3.2.3",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "pydantic-core==2.23.4",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.16",
    "pyyaml==6.0.2",
    "rich==13.9.3",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlparse==0.5.1",
    "starlette==0.41.2",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "uvicorn==0.32.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
    "supabase>=2.16.0",
]
